+++ Src/RE2.Native/Submatch.h
@@ -8,8 +8,8 @@
 #include <cstdint>
 
 struct alignas(1) Submatch {
-	uint64_t Index;
-	uint64_t Length;
+	int32_t Index;
+	int32_t Length;
 };
 
 #pragma pack(pop)
